requirements.txt
backports-abc==0.5
bleach==1.5.0
certifi==2017.1.23
configparser==3.5.0
decorator==4.0.6
deepdiff==2.5.3
Django==1.11
django-ckeditor==5.0.3
django-cors-headers==1.2.2
django-excel==0.0.4
django-extensions==1.6.7
django-fcm==0.1.1
django-model-utils==3.0.0
django-rest-multiple-models==1.8.1
django-rest-swagger==0.3.8
django-simple-history==1.8.1
django-thumbs==0.4
django-watson==1.2.1
djangorestframework==3.3.3
entrypoints==0.2.2
functools32==3.2.3.post2
future==0.15.2
graphviz==0.4.10
html5lib==0.9999999
httpie==0.9.6
ipdb==0.8.1
ipykernel==4.5.2
ipython==4.0.3
ipython-genutils==0.1.0
ipywidgets==5.2.2
jdcal==1.2
Jinja2==2.9.5
jsonfield==2.0.1
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==4.4.0
jupyter-console==5.1.0
jupyter-core==4.2.1
MarkupSafe==0.23
mistune==0.7.3
MySQL-python==1.2.5
nbconvert==5.1.1
nbformat==4.2.0
notebook==4.3.2
odfpy==0.9.6
openpyxl==2.2.2
pandocfilters==1.4.1
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.6
Pillow==3.3.0
pkg-resources==0.0.0
pluggy==0.4.0
prompt-toolkit==1.0.13
ptyprocess==0.5
py==1.4.33
pydot==1.2.2
pyexcel==0.2.0
pyexcel-io==0.1.0
pyexcel-ods==0.1.0
pyexcel-webio==0.0.6
pyexcel-xls==0.1.0
pyexcel-xlsx==0.1.0
Pygments==2.1.3
pygraphviz==1.3.1
pyparsing==2.1.10
PyPDF2==1.26.0
python-dateutil==2.6.0
pytz==2016.7
PyYAML==3.12
pyzmq==16.0.2
qtconsole==4.2.1
reportlab==3.4.0
requests==2.11.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
terminado==0.6
testpath==0.3
texttable==0.8.4
tornado==4.4.2
tox==2.7.0
traitlets==4.1.0
virtualenv==15.1.0
wcwidth==0.1.7
widgetsnbextension==1.2.6
xhtml2pdf==0.0.6
xlrd==0.9.4
XlsxWriter==0.8.4
xlwt-future==0.8.0
